Split up store
